:zap: 함수 조합 완료
