+++ backend/firstapp/templates/test.html
@@ -12,6 +12,8 @@
         <input type="text" id="purposes" name="purposes"><br>
         <label for="product">Product:</label><br>
         <input type="text" id="product" name="product"><br>
+        <label for="text">Text:</label><br>
+        <input type="text" id="texts" name="texts"><br>
         <input type="submit" value="Submit">
     </form>
 </body>
