--- backend/response_content.txt
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Example Page</title>
-</head>
-<body>
-    <h1>All TextBoxes</h1>
-    <ul>
-        
-            <li>큰 홍보문구</li>
-        
-    </ul>
-
-    <h1>All Templates</h1>
-    <ul>
-        
-            <li>1000:200</li>
-        
-            <li>1500:500</li>
-        
-    </ul>
-</body>
-</html>
