+++ docker-compose.yml
@@ -12,7 +12,7 @@ services:
     build: ./database
     # 내부에서 개방할 포트 : 외부에서 접근할 포트
     ports:
-      - "3306:3306"
+      - "3307:3306"
 
   backend:
     tty: true
